Try to update relationship attribute from csv
