+++ cadasta/gui/tools/wizard/project_update_wizard.py
@@ -21,12 +21,6 @@ from cadasta.gui.tools.wizard.step_project_update02 import (
 from cadasta.gui.tools.wizard.step_project_update03 import (
     StepProjectUpdate03
 )
-from cadasta.gui.tools.wizard.step_project_update04 import (
-    StepProjectUpdate04
-)
-from cadasta.gui.tools.wizard.step_project_update05 import (
-    StepProjectUpdate05
-)
 
 from cadasta.utilities.i18n import tr
 from cadasta.gui.tools.wizard.wizard_dialog import WizardDialog
@@ -44,7 +38,7 @@ class ProjectUpdateWizard(WizardDialog):
 
     step_project_update01 = None
     step_project_update02 = None
-    step_project_update05 = None
+    step_project_update03 = None
 
     def __init__(self, parent=None, iface=None):
         """Constructor for the dialog.
@@ -85,17 +79,17 @@ class ProjectUpdateWizard(WizardDialog):
         :return: Last step of wizard.
         :rtype: WizardStep
         """
-        return self.step_project_update05
+        return self.step_project_update03
 
     def populate_stacked_widget(self):
         """Append widgets to stacked widget."""
         self.step_project_update01 = StepProjectUpdate01(self)
         self.step_project_update02 = StepProjectUpdate02(self)
-        self.step_project_update05 = StepProjectUpdate05(self)
+        self.step_project_update03 = StepProjectUpdate03(self)
 
         self.stackedWidget.addWidget(self.step_project_update01)
         self.stackedWidget.addWidget(self.step_project_update02)
-        self.stackedWidget.addWidget(self.step_project_update05)
+        self.stackedWidget.addWidget(self.step_project_update03)
 
     def prepare_the_next_step(self, new_step):
         """Prepare the next tab.
